Make "Team" the new "Contact" section
As we don't have a contact form anylonger, make the previous "Team" section the new "Contact" section.

The "Contact" section should be, however, darker to (1) balance the overall colors of the page; and (2) emphasize the content and promote .contacting.

Changing the background is achieved by #contact.

Changing the button colors need a custom CSS class .social-buttons-dark. So, create a custom SASS file to handle customizations and extend in it the previous .social-buttons CSS class with colors corresponding to the ones used in .btn-primary.