Remove link to Scopus Author ID
This solves a potential confusion with the three links above (mobile version), as there are now only two labels.

The link to Scopus was mainly a SEO strategy, but Scopus' robots.txt actually disallows author pages, so it's not useful.

Finally, ORCID and Researcher ID are far more popular, so keep only them.