Responsive images
Use the `gulp-responsive` tool to generate smaller, optimized, and WebP versions of images from the website. To separate source and target images, move source images to the `images` directory and use `img` for the responsive versions. Add original images to `images` dir.

Use the `<picture>` HTML element and media queries to let the browser decide for the proper version.

Darken the header background image in CSS using `background-blend-mode` instead of requiring post-processing with other tools.

For the portfolio, use small and large versions instead of "full" and "thumbnail", which allow small devices to download a single "small" image for both the outer and inner visualizations.

Support for WebP images in CSS will be considered in future.