Serve JS back from the server
This reduces time to load by not requiring an extra DNS lookup + SSL connection to Cloudflare servers if resource is not cached. This also allows these resources to be server pushed when we migrate to a Cloudflare frontend, when we may also benefit back from the longer cache time directive.

This reverts
1717907cadf758285700c1db276f276d41a5c276
f2f547a913b51abc91fe065541ea355fa77cd058
266cbc98a389d4e883799ac94da5b8fd4ffddfcb